Move DigitallySignedStruct into the public API
